plugin-dev: mark symbols meant to be public with __attribute__((visibility("default")))
This is a partial backport of commit 1353b066072 ("Visibility hidden by
default"), from the master branch.

If a plugin author compiles their plugin with `-fvisibility=hidden`,
symbols such as `__bt_get_begin_section_plugin_descriptor` are local,
therefore not seen by libbbabeltrace2, and the plugin fails to load.

To avoid this, use `__attribute__((visibility("default")))` on those
symbols that we really want to be externally visible.

Reported-By: Kienan Stewart <kstewart@efficios.com>
Change-Id: Ia08173fa69ed699583a0957ce42f305d443b2952
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/11111
Reviewed-by: Michael Jeanson <mjeanson@efficios.com>
Reviewed-by: Kienan Stewart <kstewart@efficios.com>
Reviewed-by: Philippe Proulx <eeppeliteloop@gmail.com>