fix: tests: use a portable invocation of 'sed -i'
The GNU and BSD sed implementations slightly differ in the handling of
an argument to the '-i' switch to use as a suffix to a backup file. BSD
sed requires an argument while it's optional for GNU sed, if present it
must be placed directly after the switch with no space.

To make both implementations work properly and not create a backup file,
a zero length argument must be placed directly after the switch with no
space.

Change-Id: Ifd2cd5b27df05e3301e527ef1d8e718b434d2236
Signed-off-by: Michael Jeanson <mjeanson@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/11132
Reviewed-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-by: Philippe Proulx <eeppeliteloop@gmail.com>