ctf: reduce the number of `Makefile.am`s
This should help speeding parallel builds a bit, by reducing the time
spent in directories with few source files.  The time necessary to build
the intermediary .la libraries is also non-negligible (at least on
Windows), so having less of those should help.

I kept `common/metadata/Makefile.am`, because it has a lot of complexity
specific to the use of bison, I think it's good to keep it in its own
file.

Change-Id: I6b7a5ca370d6b4481bda3649ab51009e70a0e393
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/11137
Tested-by: jenkins <jenkins@lttng.org>
Reviewed-by: Philippe Proulx <eeppeliteloop@gmail.com>