tests: normalize names of files and more
The current naming situation is somewhat of a mess in the `tests`
directory, making it difficult to deduce (or write) a convention.

Use hyphens to separate words as much as possible, following what we
usually like for source file names (see `tree src`).

Pretty much the only things remaining with underscores are Python file
names because of the `import` constraints and because it's Python's own
convention.

Also make all shell scripts have the `.sh` extension. This makes any
worthy editor highlight the shell syntax correctly and tells you
immediately it's not a compiled ("binary") executable. I'd prefer
`.bash` for Bash code, but I don't think we need to go as far.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Change-Id: I43cd279d78c240a7e3686ad8afd6759bad9f575f
Reviewed-on: https://review.lttng.org/c/babeltrace/+/11170
Tested-by: jenkins <jenkins@lttng.org>
Reviewed-by: Simon Marchi <simon.marchi@efficios.com>
CI-Build: Simon Marchi <simon.marchi@efficios.com>