Sort includes in C++ files
Set the `SortIncludes` option of clang-format to `Regroup`, meaning that
it will group include statements based on the categories in
`IncludeCategories`.

In `IncludeCategories`, define categories such that the groups are (in
order of priority):

  1. C++ system headers
  2. C system headers
  3. Babeltrace 2 public headers
  4. Logging headers (logging/log.h, as well as other header files
     including it): it might be important for these files to stay at the
     top, in case there is some logging in header files, so that things
     like log tags are properly set.
  5  Internal common headers (e.g. from common, cpp-common, compat,
     logging)
  6. Plugin common files (from plugins/common)
  7. Local headers
  8. tap/tap.h: it defines macros with common names, such as `ok` and
     `fail`, which may clash with identifiers in other headers.  It's
     safer to keep it last.

Includes in files parser.ypp and lexer.lpp were sorted by hand.  Due to
the change in inclusion order, parser.hpp (generated from parser.ypp)
did not see the definition of yyscan_t anymore.  Add an inclusion of
scanner.hpp in the `%code requires` section, which ends up in both the
generated .cpp and .hpp files.

Change-Id: I0568167065b90070ac03ea4daf4d0a1927bec4b5
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/11459
Tested-by: jenkins <jenkins@lttng.org>