Add IWYU keep pragmas when including some compat headers
When editing some on Linux, IWYU and clangd report some inclusions of
`compat/endian.h`, for instance, as unused.   This is because this
header defines some macros only on systems that lack them, and Linux
isn't one of them.  These includes should not be removed, so add some
pragmas to let tools (and even humans) know that they are important.

Change-Id: Id2cba11a568567e6a75d094df3567b0847dac690
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/11518
Tested-by: jenkins <jenkins@lttng.org>